Added Confirmation to some commands!

+++ commands/Moderation/nuke.js
@@ -1,5 +1,5 @@
 const { SlashCommandBuilder } = require('@discordjs/builders');
-const { Permissions } = require('discord.js');
+const { Permissions, MessageActionRow, MessageButton } = require('discord.js');
 
 module.exports = {
   data: new SlashCommandBuilder()
@@ -10,12 +10,51 @@ module.exports = {
     const channel = interaction.channel
     // console.log(interaction.channel)
     // interaction.reply({ content: 'Check the console!', ephemeral: true })
-    await interaction.reply("Nuking channel...")
-    await interaction.channel.clone()
-    .then(async (cnl) => {
-      await cnl.setParent(channel.parent.id)
-      await cnl.setPosition(channel.position)
-      await channel.delete()
-  })
+    const row = new MessageActionRow()
+      .addComponents(
+        new MessageButton()
+          .setStyle("SUCCESS")
+          .setLabel("YES")
+          .setCustomId("nuke-yes")
+          .setEmoji("💥"),
+        new MessageButton()
+          .setStyle("DANGER")
+          .setLabel("NO")
+          .setCustomId("nuke-no")
+          .setEmoji("❌"),
+      )
+    await interaction.reply({
+      content: 'Are you sure you want to nuke this channel?\nConfirmation will be cancelled after 30 seconds.',
+      ephemeral: true,
+      components: [
+        row
+      ]
+    })
+
+    const filter = (btnInt) => {
+      return interaction.user.id == btnInt.user.id
+    }
+
+    const collector = channel.createMessageComponentCollector({
+      filter,
+      max: 1,
+      time: 1000 * 30
+    })
+
+    collector.on('end', async (collection) => {
+      if(collection.first()?.customId == 'nuke-yes') {
+        await interaction.editReply({ content: 'Nuking channel...', components: [], ephemeral: true })
+        await interaction.channel.clone()
+        .then(async (cnl) => {
+          await cnl.setParent(channel.parent.id)
+          await cnl.setPosition(channel.position)
+          await channel.delete()
+        })
+      } else if(collection.first()?.customId == 'nuke-no') {
+        return await interaction.editReply({ content: "The nuke has been disarmed.", components: [], ephemeral: true })
+      } else {
+        return await interaction.editReply({ content: "You didn't respond in time.", components: [], ephemeral: true })
+      }
+    })
  }
 }
